Move `PParams` local state queries to separate module
