+++ lib/network-layer/cardano-wallet-network-layer.cabal
@@ -48,6 +48,7 @@ library
     Cardano.Wallet.Network.Implementation.UnliftIO
     Cardano.Wallet.Network.Light
     Cardano.Wallet.Network.LocalStateQuery.Extra
+    Cardano.Wallet.Network.LocalStateQuery.PParams
     Cardano.Wallet.Network.LocalStateQuery.RewardAccount
     Cardano.Wallet.Network.LocalStateQuery.StakeDistribution
     Cardano.Wallet.Network.Logging
